Top画面の遷移機能を１つのControllerへまとめる
